+++ core/kscreenmanager.py
@@ -1,10 +1,17 @@
-from kivy.uix.screenmanager import ScreenManager, NoTransition
+from kivy.uix.screenmanager import Screen, ScreenManager, CardTransition
 from .base.kobject import KObject
 
 class KScreenManager(ScreenManager, KObject):
 
   def __init__(self, **kwargs):
     super(KScreenManager, self).__init__(**kwargs)
-    self.transition = NoTransition()
+    self.transition = CardTransition()
+
+  pass
+
+class KScreen(Screen, KObject):
+
+  def __init__(self, **kwargs):
+    super(KScreen, self).__init__(**kwargs)
 
   pass
