+++ screens/pages/login_screen.kv
@@ -31,12 +31,11 @@
       write_tab: False
 
     LoginButton:
-      id: login_button
       emp_code_inp: emp_code
       emp_pass_inp: emp_pass
       text: "LOGIN"
       center_x: root.center_x
       top: root.top - 230
-      size_hint: None, None
+      size_hint: [None, None]
       width: 300 + (self.ids.lbl_txt.texture_size[0] - self.ids.lbl_txt.texture_size[0])
       height: 30
