+++ screens/pages/queue_screen.kv
@@ -14,6 +14,7 @@
       pos_hint: {"center_y": .5}
       on_press:
         root.app.mainscreenmanager.current = 'session_screen'
+        root.app.cameraimage.open_camera()
 
   MDCardSwipeFrontBox:
