+++ screens/pages/session_screen.kv
@@ -4,15 +4,24 @@
   BoxLayout:
     orientation: "vertical"
 
-    MDLabel:
-
-      font_size: 35
+    BoxLayout:
       size_hint: [None, 0.1]
       width: root.width - 10
-      pos_hint: {"center_x": 0.5}
-      text_size: (None, None)
-      markup: True
-      text: "[/color][color=#607D8B]ESMS[/color][color=#888888][size=28] Workspace[/size]"
+
+      MDLabel:
+        size_hint: [0.9, 1]
+        font_size: 35
+        pos_hint: {"center_x": 0.5}
+        text_size: (None, None)
+        markup: True
+        text: "[/color][color=#607D8B]ESMS[/color][color=#888888][size=28] Workspace[/size]"
+
+      BoxLayout:
+        size_hint: [0.1, 0.3]
+        pos_hint: {"center_y": 0.5}
+
+        CompleteSessionButton:
+          text: "COMPLETE SESSION"
 
     BoxLayout:
       spacing: 5
@@ -21,90 +30,70 @@
       pos_hint: {"center_x": 0.5}
 
       BoxLayout:
+        spacing: 5
+        orientation: "vertical"
+        size_hint: [0.2, 1]
 
         BoxLayout:
-          spacing: 5
-          orientation: "vertical"
-          size_hint: [0.5, 1]
+          size_hint: [1, 0.3]
+
+          canvas.before:
+            Color:
+              rgba: hex('#607D8B')
+            Line:
+              width: 0.5
+              rectangle: self.x, self.y, self.width, self.height
 
           BoxLayout:
-            size_hint: [1, 0.5]
-
-            BoxLayout:
-
-              canvas.before:
-                Color:
-                  rgba: hex('#607D8B')
-                Line:
-                  width: 0.5
-                  rectangle: self.x, self.y, self.width, self.height
+            size_hint: [0.05, 1]
 
           BoxLayout:
-            size_hint: [1, 0.5]
+            orientation: "vertical"
+            size_hint: [0.9, 0.9]
+            pos_hint: {"center_y": 0.5}
 
-            BoxLayout:
+            canvas.before:
+              Color:
+                rgba: hex(app.emotion_color)
+              Line:
+                width: 1.5
+                rectangle: self.x - 1, self.y - 3, self.width + 2, self.height + 6
 
-              canvas.before:
-                Color:
-                  rgba: hex('#607D8B')
-                Line:
-                  width: 0.5
-                  rectangle: self.x, self.y, self.width, self.height
+            # BoxLayout:
+            #   size_hint: [1, None]
+            #   height: "2dp"
+
+            CameraImage:
+              pos_hint: {"center_x": 0.5, "center_y": 0.5}
+              source: 'assets/video.jpg'
+              nocache: True
+
+            # BoxLayout:
+            #   size_hint: [1, None]
+            #   height: "2dp"
+
+          BoxLayout:
+            size_hint: [0.05, 1]
+
+        BoxLayout:
+          size_hint: [1, 0.7]
+
+          canvas.before:
+            Color:
+              rgba: hex('#607D8B')
+            Line:
+              width: 0.5
+              rectangle: self.x, self.y, self.width, self.height
 
       BoxLayout:
+        size_hint: [0.8, 1]
 
-        BoxLayout:
-          spacing: 5
-          orientation: "vertical"
-          size_hint: [0.5, 1]
-
-          BoxLayout:
-            size_hint: [1, 0.5]
-
-            BoxLayout:
-
-              canvas.before:
-                Color:
-                  rgba: hex('#607D8B')
-                Line:
-                  width: 0.5
-                  rectangle: self.x, self.y, self.width, self.height
-
-          BoxLayout:
-            size_hint: [1, 0.5]
-
-            BoxLayout:
-
-              canvas.before:
-                Color:
-                  rgba: hex('#607D8B')
-                Line:
-                  width: 0.5
-                  rectangle: self.x, self.y, self.width, self.height
+        canvas.before:
+          Color:
+            rgba: hex('#607D8B')
+          Line:
+            width: 0.5
+            rectangle: self.x, self.y, self.width, self.height
 
     BoxLayout:
       size_hint: [1, 0.01]
-
-  # BoxLayout:
-  #   orientation: 'vertical'
-  #   padding: root.width * 0.05, root.height * 0.05
-  #   spacing: 5
-
-  #   BoxLayout:
-  #     size_hint: [1, 0.75]
-  #     Image:
-  #       id: image_source
-  #       source: 'assets/video.jpg'
-  #       size_hint_y: None
-  #       height: 200
-  #       nocache: True
-
-  #   BoxLayout:
-  #     size_hint: [1, 0.1]
-  #     SessionButton:
-  #       image_source: image_source
-  #       text: "Start SESSION"
-  #       center_x: root.center_x
-  #       size_hint: None, None
-  #       width: root.width * 0.9 + (self.ids.lbl_txt.texture_size[0] - self.ids.lbl_txt.texture_size[0])
-  #       height: 30
