+++ widgets/widget_list.py
@@ -4,4 +4,5 @@ from widgets.login_screen.login_button import LoginButton
 from widgets.queue_screen.queue_item import QueueItem
 from widgets.queue_screen.queue_list import QueueList
 ## Session Screen widgets
-from widgets.session_screen.session_button import SessionButton
+from widgets.session_screen.camera_image import CameraImage
+from widgets.session_screen.session_button import CompleteSessionButton
